Moved hearing request type logic to concern
